Optimize inventory and complaints pipeline performance
- Pre-map StoreKey to WarehouseKey in sales workers so inventory micro-aggregation operates at warehouse grain (27 groups) instead of store grain (500 groups), eliminating the expensive post-hoc rollup that was taking ~99s on 200M sales rows
- Remove unnecessary drop_duplicates in complaints accumulator since SalesOrderNumbers are unique per chunk by construction, saving ~110s on 200M sales rows
- Add warning when unmapped stores are dropped during inventory micro-aggregation for data integrity visibility
- Fix misleading complexity comment in wishlists accumulator with accurate explanation of why cross-chunk dedup is needed

+++ src/dimensions/employees/transfers.py
@@ -338,7 +338,7 @@ def apply_transfers(
     # Sort and reassign keys/sequences
     df = df.sort_values(["EmployeeKey", "StartDate"]).reset_index(drop=True)
     df["AssignmentKey"] = np.arange(1, len(df) + 1, dtype=np.int32)
-    df["AssignmentSequence"] = df.groupby("EmployeeKey").cumcount().astype(np.int32) + 1
+    df["AssignmentSequence"] = (df.groupby("EmployeeKey").cumcount() + 1).astype(np.int16)
 
     # Normalize dates
     df["StartDate"] = pd.to_datetime(df["StartDate"]).dt.normalize()
